Rename base.py to base_backend.py for clarity
